adjust subvolumes for snapper use
For each subvolume snapper is configured to work against, it creates a
folder `$subvol/.snapshots`. We can remount the original @snapshots
subvol onto both mount points, but it would cause snapper write
conflicts since it just names each snapshot as a number. So if 2
subvolumes both have their first snapshot and get saved as `1`, you get
the idea. So making another subvolume for storing snapshots of home.

Snapper also gets unhappy when it goes to create the config if the
`.snapshots` folder already exists, so let's leave that for userspace
configuration. Thankfully, it's easy enough:

1. `snapper -c $name create-config /path/to/subvol`
2. edit `/etc/fstab` to mount the new subvol similar to the others
3. `mount -a` to ensure they're all mounted