add ability to format system disk with random data before encryption
If user says yes, the script will:
- create an initial LUKS container with a fake key and use that to
  encrypt the system disk
- write from `/dev/zero` to the encrypted container to fill it with
  random data
- overwrite the LUKS header on the fake device to blow away the temp key
- unmount the fake device
- create a new (real/final) LUKS container that will be used for the
  rest of the installation

Tested working, though `dd` does show the expected `No space left on
device` error writing to the fake device because we just tell it to fill
and don't provide a block count or anything. Could redirect stderr to
`/dev/null`, but it doesn't bother me that much and I'd rather have it
printed if I ever have to debug it.